+++ unique_toolkit/tests/language_model/test_language_models_service_unit.py
@@ -3,10 +3,14 @@ from unittest.mock import patch
 
 import pytest
 import unique_sdk
+from pydantic import BaseModel
 
 from tests.test_obj_factory import get_event_obj
 from unique_toolkit.content.schemas import ContentChunk
-from unique_toolkit.language_model.functions import _prepare_completion_params_util
+from unique_toolkit.language_model.functions import (
+    _add_response_format_to_options,
+    _prepare_completion_params_util,
+)
 from unique_toolkit.language_model.infos import LanguageModelName
 from unique_toolkit.language_model.schemas import (
     LanguageModelMessage,
@@ -39,6 +43,11 @@ mock_tool = LanguageModelTool(
         required=["location"],
     ),
 )
+
+
+class PydanticModel(BaseModel):
+    name: str
+    age: int
 
 
 class TestLanguageModelServiceUnit:
@@ -698,3 +707,31 @@ class TestLanguageModelServiceUnit:
             },
             startText=None,
         )
+
+    def test_add_output_schema_from_pydantic_enforce_schema(self):
+        options = {}
+        options = _add_response_format_to_options(
+            options, PydanticModel, structured_output_enforce_schema=True
+        )
+        assert options["responseFormat"] == {
+            "type": "json_schema",
+            "json_schema": {
+                "name": "PydanticModel",
+                "strict": True,
+                "schema": PydanticModel.model_json_schema(),
+            },
+        }
+
+    def test_add_output_schema_from_pydantic_no_enforce_schema(self):
+        options = {}
+        options = _add_response_format_to_options(
+            options, PydanticModel, structured_output_enforce_schema=False
+        )
+        assert options["responseFormat"] == {
+            "type": "json_schema",
+            "json_schema": {
+                "name": "PydanticModel",
+                "strict": False,
+                "schema": PydanticModel.model_json_schema(),
+            },
+        }