+++ unique_toolkit/unique_toolkit/language_model/functions.py
@@ -1,7 +1,8 @@
 import logging
-from typing import Optional, cast
+from typing import Optional, Type, cast
 
 import unique_sdk
+from pydantic import BaseModel
 
 from unique_toolkit.content.schemas import ContentChunk
 from unique_toolkit.evaluators import DOMAIN_NAME
@@ -28,6 +29,8 @@ def complete(
     timeout: int = DEFAULT_COMPLETE_TIMEOUT,
     tools: Optional[list[LanguageModelTool]] = None,
     other_options: Optional[dict] = None,
+    structured_output_model: Optional[Type[BaseModel]] = None,
+    structured_output_enforce_schema: bool = False,
 ) -> LanguageModelResponse:
     """
     Calls the completion endpoint synchronously without streaming the response.
@@ -50,6 +53,8 @@ def complete(
         temperature=temperature,
         tools=tools,
         other_options=other_options,
+        structured_output_model=structured_output_model,
+        structured_output_enforce_schema=structured_output_enforce_schema,
     )
 
     try:
@@ -77,6 +82,8 @@ async def complete_async(
     timeout: int = DEFAULT_COMPLETE_TIMEOUT,
     tools: Optional[list[LanguageModelTool]] = None,
     other_options: Optional[dict] = None,
+    structured_output_model: Optional[Type[BaseModel]] = None,
+    structured_output_enforce_schema: bool = False,
 ) -> LanguageModelResponse:
     """
     Calls the completion endpoint asynchronously without streaming the response.
@@ -106,6 +113,8 @@ async def complete_async(
         temperature=temperature,
         tools=tools,
         other_options=other_options,
+        structured_output_model=structured_output_model,
+        structured_output_enforce_schema=structured_output_enforce_schema,
     )
 
     try:
@@ -292,6 +301,22 @@ def _to_search_context(chunks: list[ContentChunk]) -> dict | None:
     ]
 
 
+def _add_response_format_to_options(
+    options: dict,
+    structured_output_model: Type[BaseModel],
+    structured_output_enforce_schema: bool = False,
+) -> dict:
+    options["responseFormat"] = {
+        "type": "json_schema",
+        "json_schema": {
+            "name": structured_output_model.__name__,
+            "strict": structured_output_enforce_schema,
+            "schema": structured_output_model.model_json_schema(),
+        },
+    }
+    return options
+
+
 def _prepare_completion_params_util(
     messages: LanguageModelMessages,
     model_name: LanguageModelName | str,
@@ -299,6 +324,8 @@ def _prepare_completion_params_util(
     tools: Optional[list[LanguageModelTool]] = None,
     other_options: Optional[dict] = None,
     content_chunks: Optional[list[ContentChunk]] = None,
+    structured_output_model: Optional[Type[BaseModel]] = None,
+    structured_output_enforce_schema: bool = False,
 ) -> tuple[dict, str, dict, Optional[dict]]:
     """
     Prepares common parameters for completion requests.
@@ -312,6 +339,10 @@ def _prepare_completion_params_util(
     """
 
     options = _add_tools_to_options({}, tools)
+    if structured_output_model:
+        options = _add_response_format_to_options(
+            options, structured_output_model, structured_output_enforce_schema
+        )
     options["temperature"] = temperature
     if other_options:
         options.update(other_options)